# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Lumen",
        "description": "This is Documentation of Lumen",
        "contact": {
            "email": "brayenprayoga41@gmail.com"
        },
        "version": "3.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/",
            "description": "Localhost"
        }
    ],
    "paths": {
        "/auth/login": {
            "post": {
                "tags": [
                    "Login"
                ],
                "operationId": "8cb7742ae98d0990acb2907c4e00bf24",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "title": "email",
                                        "description": "email",
                                        "example": "no-reply@gmail.com"
                                    },
                                    "password": {
                                        "title": "password",
                                        "description": "Password",
                                        "example": "P@ssw0rd"
                                    },
                                    "role": {
                                        "title": "Role",
                                        "description": "Role",
                                        "example": "DASHBOARD/SELF_SERVICE/COUNTER"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "tags": [
                    "Register"
                ],
                "operationId": "ed42f392010c301892e640067141676e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nama": {
                                        "title": "nama",
                                        "description": "nama",
                                        "example": ""
                                    },
                                    "email": {
                                        "title": "email",
                                        "description": "email",
                                        "example": "no-reply@gmail.com"
                                    },
                                    "password": {
                                        "title": "password",
                                        "description": "Password",
                                        "example": "P@ssw0rd"
                                    },
                                    "role": {
                                        "title": "Role",
                                        "description": "Role",
                                        "example": "DASHBOARD/SELF_SERVICE/COUNTER"
                                    },
                                    "kode_dealer": {
                                        "title": "Kode Dealer",
                                        "description": "Kode Dealer",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/user/master-dealer": {
            "get": {
                "tags": [
                    "Master Dealer"
                ],
                "operationId": "922f31d8757b16ba877a9f3a02b2c3ea",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/user/master-dealer/simpan": {
            "post": {
                "tags": [
                    "Master Dealer"
                ],
                "operationId": "1a8c050c2341c476a9a05fbc3a8b241d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "kode_dealer": {
                                        "title": "Kode Dealer",
                                        "description": "Kode Dealer",
                                        "example": ""
                                    },
                                    "nama": {
                                        "title": "nama",
                                        "description": "nama",
                                        "example": ""
                                    },
                                    "kode_menu": {
                                        "title": "Kode Menu",
                                        "description": "Kode Menu",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/user/master-dealer/edit": {
            "post": {
                "tags": [
                    "Master Dealer"
                ],
                "operationId": "51c4dd80910d12255d3888f7249af211",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "kode_dealer": {
                                        "title": "Kode Dealer",
                                        "description": "Kode Dealer",
                                        "example": ""
                                    },
                                    "nama": {
                                        "title": "nama",
                                        "description": "nama",
                                        "example": ""
                                    },
                                    "kode_menu": {
                                        "title": "Kode Menu",
                                        "description": "Kode Menu",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/user/master-dealer/hapus": {
            "get": {
                "tags": [
                    "Master Dealer"
                ],
                "operationId": "934429793f7ada1b9dc3124703f92b11",
                "parameters": [
                    {
                        "name": "kode_dealer",
                        "in": "query",
                        "description": "Kode Dealer",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Hubungi Developer Untuk Akun Login",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "Passport",
                "scheme": "bearer"
            }
        }
    }
}